refactor: extract load_diagnostics_table_for_root into beamtalk-core
Both beamtalk-compiler-port and beamtalk-lsp independently implemented
the same per-root "read beamtalk.toml → parse [diagnostics] → return
empty on failure" pattern. The compiler-port even commented that it was
"mirroring the LSP's load_diagnostics_table". Add a shared
`load_diagnostics_table_for_root` function to
`beamtalk-core::compilation::diagnostics_policy` (where
`parse_diagnostics_table_from_manifest_toml` already lives for this
same cross-crate sharing reason, BT-2800), export it from the
`compilation` module, and update both callers to delegate to it.

Behaviour is identical: missing manifest → empty table (no-op),
parse error → WARN log + empty table. Caller-specific telemetry
(the compiler-port's debug log on success, the LSP's merge-conflict
warn) stays in each caller.
