+++ crates/beamtalk-core/src/compilation/mod.rs
@@ -18,8 +18,8 @@ pub mod extension_index;
 pub use dependency::{DependencyMap, DependencySource, DependencySpec, GitReference};
 pub use diagnostics_policy::{
     DiagnosticSeverityOverride, DiagnosticsTable, DiagnosticsTableError, apply_diagnostics_table,
-    parse_diagnostics_table, parse_diagnostics_table_from_manifest_toml,
-    parse_package_name_from_manifest_toml,
+    load_diagnostics_table_for_root, parse_diagnostics_table,
+    parse_diagnostics_table_from_manifest_toml, parse_package_name_from_manifest_toml,
 };
 pub use extension_conflicts::{
     ExtensionConflict, conflict_diagnostics, detect_extension_conflicts, shadowing_diagnostics,
