+++ crates/beamtalk-lsp/src/server.rs
@@ -809,43 +809,29 @@ impl Backend {
     /// `beamtalk.toml` edits made while the server is running require an LSP
     /// restart to take effect.
     async fn load_diagnostics_table(&self, roots: &[PathBuf]) {
-        use beamtalk_core::compilation::diagnostics_policy::parse_diagnostics_table_from_manifest_toml;
-
         let roots_owned: Vec<PathBuf> = roots.to_vec();
         let table = tokio::task::spawn_blocking(move || {
             let mut merged = beamtalk_core::compilation::DiagnosticsTable::new();
             for root in &roots_owned {
-                let manifest_path = root.join("beamtalk.toml");
-                let Ok(content) = std::fs::read_to_string(&manifest_path) else {
-                    continue;
-                };
-                match parse_diagnostics_table_from_manifest_toml(&content) {
-                    Ok(root_table) => {
-                        for (category, severity) in root_table {
-                            if let Some(previous) = merged.get(&category) {
-                                if *previous != severity {
-                                    tracing::warn!(
-                                        root = %root.display(),
-                                        category = ?category,
-                                        previous = ?previous,
-                                        new = ?severity,
-                                        "[diagnostics] category set differently by multiple \
-                                         workspace roots; this root's value wins for the \
-                                         whole session"
-                                    );
-                                }
-                            }
-                            merged.insert(category, severity);
+                // Missing or unreadable manifest → empty table → no-op merge.
+                // Parse errors are logged inside load_diagnostics_table_for_root.
+                let root_table =
+                    beamtalk_core::compilation::load_diagnostics_table_for_root(root);
+                for (category, severity) in root_table {
+                    if let Some(previous) = merged.get(&category) {
+                        if *previous != severity {
+                            tracing::warn!(
+                                root = %root.display(),
+                                category = ?category,
+                                previous = ?previous,
+                                new = ?severity,
+                                "[diagnostics] category set differently by multiple \
+                                 workspace roots; this root's value wins for the \
+                                 whole session"
+                            );
                         }
                     }
-                    Err(e) => {
-                        tracing::warn!(
-                            path = %manifest_path.display(),
-                            error = %e,
-                            "failed to parse [diagnostics] table in beamtalk.toml; \
-                             using Rule 1 defaults for this root"
-                        );
-                    }
+                    merged.insert(category, severity);
                 }
             }
             merged
